requirements.txt
# CLaRa-inspired Legal Document Analysis System
# Mississippi Secretary of State - AI Innovation Hub Phase 1

# Core dependencies
anthropic>=0.40.0
openai>=1.50.0

# Document processing
pypdf>=4.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
typer>=0.12.0

# Async support
aiofiles>=24.0.0
